Tetris game
For my Tetris game, I created an interface to run as an applet that can hold a 10 * 20 grid of filled squares. And I add the Tetris object Block at a JPanel to my TetrisApplet left and the get the main view for my Tetris game from TetrisView class. I draw the game background at black and random all colours which I already define in the colors array without black colour for my Tetris squares. All of this, I add in my main JApplet fire; there are also have an extend view to look for next going down square when my game is started. Moreover, I add my name and registration number in the title of the applet.
The basic structure of my program is a 10 * 20 grid draw at the TetrisView class, and then adds the main view in my main Tetris object block. The Panel block is a main control object which controls each square moving though mouse button respectively. The left mouse button is make the squares can moving left and the right mouse button is make the squares moving right. And the button which in the mi